feat: implement tasks CLI and modules
