+++ src/cli/args.js
@@ -1,5 +1,11 @@
 const parseArgs = () => {
-    // Write your code here 
+  const parseArgsArr = process.argv.slice(2);
+
+  for (let i = 0; i < parseArgsArr.length; i += 2) {
+    const key = parseArgsArr[i].slice(2);
+    const value = parseArgsArr[i + 1];
+    console.log(`${key} is ${value}`);
+  }
 };
 
 parseArgs();
