+++ src/cli/env.js
@@ -1,5 +1,10 @@
 const parseEnv = () => {
-    // Write your code here 
+  const prefix = "RSS_";
+
+  const result = Object.entries(process.env)
+    .filter(([key]) => key.startsWith(prefix))
+    .map(([key, value]) => `${key}=${value}`)
+    .join("; ");
 };
 
 parseEnv();
